pyproject.toml
[tool.poetry]
name = "vermax"
version = "0.1.0"
description = "A simple and secure \"Home broker API\" for managing user accounts, deposits, balances, and asset purchases."
authors = ["Vinicius Reis <v_reis@outlook.com.br>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.6"
python-decouple = "^3.8"
witch-doctor = "^1.2.0"
pyfiglet = "^1.0.2"
loguru = "^0.7.2"
uvicorn = "^0.32.1"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.36"}
alembic = "^1.14.0"
asyncpg = "^0.30.0"
pydantic = {extras = ["email"], version = "^2.10.3"}
pwdlib = {extras = ["argon2"], version = "^0.2.1"}
pyjwt = "^2.10.1"
python-multipart = "^0.0.19"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-asyncio = "^0.24.0"
pytest-cov = "^6.0.0"
ruff = "^0.8.1"

[tool.pytest.ini_options]
pythonpath = "."
addopts = "--cov=src/application/use_cases --cov-report=term-missing"
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.lint.per-file-ignores]
"cpf_validator.py" = ["PLR2004"]
"user_request.py" = ["PLR2004"]
"witch_doctor_container_config_infrastructure.py" = ["E501"]

[tool.ruff.format]
preview = true
quote-style = 'single'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
